Define database

+++ core_project/src/main/java/com/an/core_project/domain/ProjectRepository.kt
@@ -1,7 +1,7 @@
 package com.an.core_project.domain
 
 interface ProjectRepository {
-    fun load(id: Int): Project
+    fun load(id: Long): Project
     fun saveCurrent()
     fun loadThumbnails(): List<ProjectSummary>
